+++ tower-http/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ServeDir` and `ServeFile`: Ability to serve precompressed files ([#156])
 - `Trace`: Add `DefaultMakeSpan::level` to make log level of tracing spans easily configurable ([#124])
+- Change the response body error type of `Compression` and `Decompression` to
+  `Box<dyn std::error::Error + Send + Sync>`. This makes them usable if the body
+  they're wrapping uses `Box<dyn std::error::Error + Send + Sync>` as its error
+  type which they previously weren't
+- Remove `BodyOrIoError`. Its been replaced with `Box<dyn std::error::Error +
+  Send + Sync>`
 
 [#124]: https://github.com/tower-rs/tower-http/pull/124
 [#156]: https://github.com/tower-rs/tower-http/pull/156